비밀번호 안보이게 처리 및 사용자 삭제 로직 수정

See merge request merge-development/backend-sg/pumpkin!21

+++ src/domains/users/users.service.ts
@@ -111,6 +111,10 @@ export class UsersService {
 
   // 특정 유저 삭제
   deleteUser(userId: string): Promise<void> {
+    if (!userId) {
+      throw new NotFoundException(USER_EXCEPTION.USER_NOT_FOUND);
+    }
+
     return this.userRepository.deleteUser(userId);
   }
